chore: DWH restructuring
