+++ insonmnia/dwh/server.go
@@ -3,16 +3,10 @@ package dwh
 import (
 	"crypto/ecdsa"
 	"database/sql"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"math/big"
 	"net"
-	"reflect"
 	"sync"
-	"time"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/grpc-ecosystem/go-grpc-prometheus"
 	_ "github.com/mattn/go-sqlite3"
 	log "github.com/noxiouz/zapctx/ctxlog"
@@ -31,14 +25,14 @@ import (
 )
 
 type DWH struct {
+	logger      *zap.Logger
 	mu          sync.RWMutex
 	ctx         context.Context
-	cfg         *Config
+	cfg         *DWHConfig
 	key         *ecdsa.PrivateKey
 	cancel      context.CancelFunc
 	grpc        *grpc.Server
 	http        *rest.Server
-	logger      *zap.Logger
 	db          *sql.DB
 	creds       credentials.TransportCredentials
 	certRotator util.HitlessCertRotator
@@ -47,7 +41,7 @@ type DWH struct {
 	lastEvent   *blockchain.Event
 }
 
-func NewDWH(ctx context.Context, cfg *Config, key *ecdsa.PrivateKey) (*DWH, error) {
+func NewDWH(ctx context.Context, cfg *DWHConfig, key *ecdsa.PrivateKey) (*DWH, error) {
 	ctx, cancel := context.WithCancel(ctx)
 	w := &DWH{
 		ctx:    ctx,
@@ -75,21 +69,9 @@ func (m *DWH) Serve() error {
 	}
 	m.blockchain = bch
 
-	if err := m.setupDB(); err != nil {
+	if m.storage, err = setupDB(m.ctx, m.db, m.blockchain); err != nil {
 		m.Stop()
 		return fmt.Errorf("failed to setupDB: %v", err)
-	}
-
-	go m.monitorBlockchain()
-	if m.cfg.ColdStart {
-		if err := m.coldStart(); err != nil {
-			m.Stop()
-			return fmt.Errorf("failed to coldStart: %v", err)
-		}
-	} else {
-		if err := m.storage.CreateIndices(m.db); err != nil {
-			return fmt.Errorf("failed to CreateIndices (Serve): %v", err)
-		}
 	}
 
 	wg := errgroup.Group{}
@@ -119,26 +101,6 @@ func (m *DWH) stop() {
 	if m.http != nil {
 		m.http.Close()
 	}
-}
-
-func (m *DWH) setupDB() error {
-	numBenchmarks, err := m.blockchain.Market().GetNumBenchmarks(m.ctx)
-	if err != nil {
-		m.stop()
-		return fmt.Errorf("failed to GetNumBenchmarks: %v", err)
-	}
-	if numBenchmarks >= NumMaxBenchmarks {
-		m.stop()
-		return errors.New("market number of benchmarks is greater than NumMaxBenchmarks")
-	}
-
-	var storage = newPostgresStorage(numBenchmarks)
-	if err := storage.Setup(m.db); err != nil {
-		return fmt.Errorf("failed to setup storage: %v", err)
-	}
-
-	m.storage = storage
-	return nil
 }
 
 func (m *DWH) serveGRPC() error {
@@ -384,1020 +346,4 @@ func (m *DWH) GetWorkers(ctx context.Context, request *pb.WorkersRequest) (*pb.W
 	}
 
 	return &pb.WorkersReply{Workers: workers, Count: count}, nil
-}
-
-func (m *DWH) monitorBlockchain() error {
-	m.logger.Info("starting monitoring")
-
-	for {
-		select {
-		case <-m.ctx.Done():
-			m.logger.Info("context cancelled (monitorBlockchain)")
-			return nil
-		default:
-			if err := m.watchMarketEvents(); err != nil {
-				m.logger.Warn("failed to watch market events, retrying", zap.Error(err))
-			}
-		}
-	}
-}
-
-func (m *DWH) watchMarketEvents() error {
-	var err error
-	m.lastEvent, err = m.getLastEvent()
-	if err != nil {
-		m.lastEvent = &blockchain.Event{}
-		if err := m.insertLastEvent(m.lastEvent); err != nil {
-			return err
-		}
-	}
-
-	m.logger.Info("starting from block", zap.Uint64("block_number", m.lastEvent.BlockNumber))
-	filter := m.blockchain.Events().GetMarketFilter(big.NewInt(0).SetUint64(m.lastEvent.BlockNumber))
-	events, err := m.blockchain.Events().GetEvents(m.ctx, filter)
-	if err != nil {
-		return err
-	}
-
-	var (
-		eventsCount int
-		dispatcher  = newEventDispatcher(m.logger)
-		tk          = time.NewTicker(time.Millisecond * 500)
-	)
-	defer tk.Stop()
-
-	// Store events by their type, run events of each type in parallel after a timeout
-	// or after a certain number of events is accumulated.
-	for {
-		select {
-		case <-m.ctx.Done():
-			m.logger.Info("context cancelled (watchMarketEvents)")
-			return nil
-		case <-tk.C:
-			m.processEvents(dispatcher)
-			eventsCount, dispatcher = 0, newEventDispatcher(m.logger)
-		case event, ok := <-events:
-			if !ok {
-				return errors.New("events channel closed")
-			}
-
-			if event.PrecedesOrEquals(m.lastEvent) {
-				continue
-			}
-
-			dispatcher.Add(event)
-			m.lastEvent, eventsCount = event, eventsCount+1
-			if eventsCount >= m.cfg.NumWorkers {
-				m.processEvents(dispatcher)
-				eventsCount, dispatcher = 0, newEventDispatcher(m.logger)
-			}
-		}
-	}
-}
-
-func (m *DWH) processEvents(dispatcher *eventsDispatcher) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	m.processEventsSynchronous(dispatcher.NumBenchmarksUpdated)
-	m.processEventsSynchronous(dispatcher.WorkersAnnounced)
-	m.processEventsSynchronous(dispatcher.WorkersConfirmed)
-	m.processEventsSynchronous(dispatcher.ValidatorsCreated)
-	m.processEventsSynchronous(dispatcher.CertificatesCreated)
-	m.processEventsAsync(dispatcher.OrdersOpened)
-	m.processEventsAsync(dispatcher.DealsOpened)
-	m.processEventsSynchronous(dispatcher.DealChangeRequestsSent)
-	m.processEventsAsync(dispatcher.Billed)
-	m.processEventsAsync(dispatcher.DealChangeRequestsUpdated)
-	m.processEventsAsync(dispatcher.OrdersClosed)
-	m.processEventsAsync(dispatcher.DealsClosed)
-	m.processEventsAsync(dispatcher.ValidatorsDeleted)
-	m.processEventsAsync(dispatcher.AddedToBlacklist)
-	m.processEventsAsync(dispatcher.RemovedFromBlacklist)
-	m.processEventsAsync(dispatcher.WorkersRemoved)
-	m.processEventsAsync(dispatcher.Other)
-
-	m.saveLastEvent()
-}
-
-func (m *DWH) processEventsSynchronous(events []*blockchain.Event) {
-	for _, event := range events {
-		m.processEventWithRetries(event)
-	}
-}
-
-func (m *DWH) processEventsAsync(events []*blockchain.Event) {
-	wg := &sync.WaitGroup{}
-	for _, event := range events {
-		wg.Add(1)
-		go func(wg *sync.WaitGroup, event *blockchain.Event) {
-			defer wg.Done()
-			if err := m.processEventWithRetries(event); err != nil {
-				m.logger.Warn("failed to processEvent, STATE IS INCONSISTENT", zap.Error(err),
-					zap.Uint64("block_number", event.BlockNumber),
-					zap.String("event_type", reflect.TypeOf(event.Data).String()),
-					zap.Any("event_data", event.Data))
-			}
-		}(wg, event)
-	}
-	wg.Wait()
-}
-
-func (m *DWH) processEventWithRetries(event *blockchain.Event) error {
-	var (
-		err        error
-		numRetries = 60
-	)
-	for numRetries > 0 {
-		if err = m.processEvent(event); err != nil {
-			m.logger.Warn("failed to processEvent, retrying", zap.Error(err),
-				zap.Uint64("block_number", event.BlockNumber),
-				zap.String("event_type", reflect.TypeOf(event.Data).String()),
-				zap.Any("event_data", event.Data))
-		} else {
-			m.logger.Debug("processed event", zap.Uint64("block_number", event.BlockNumber),
-				zap.String("event_type", reflect.TypeOf(event.Data).String()),
-				zap.Any("event_data", event.Data))
-			return nil
-		}
-		numRetries--
-		time.Sleep(time.Second)
-	}
-	return err
-}
-
-func (m *DWH) processEvent(event *blockchain.Event) error {
-	switch value := event.Data.(type) {
-	case *blockchain.NumBenchmarksUpdatedData:
-		return m.onNumBenchmarksUpdated(value.NumBenchmarks)
-	case *blockchain.DealOpenedData:
-		return m.onDealOpened(value.ID)
-	case *blockchain.DealUpdatedData:
-		return m.onDealUpdated(value.ID)
-	case *blockchain.OrderPlacedData:
-		return m.onOrderPlaced(event.TS, value.ID)
-	case *blockchain.OrderUpdatedData:
-		return m.onOrderUpdated(value.ID)
-	case *blockchain.DealChangeRequestSentData:
-		return m.onDealChangeRequestSent(event.TS, value.ID)
-	case *blockchain.DealChangeRequestUpdatedData:
-		return m.onDealChangeRequestUpdated(event.TS, value.ID)
-	case *blockchain.BilledData:
-		return m.onBilled(event.TS, value.DealID, value.PaidAmount)
-	case *blockchain.WorkerAnnouncedData:
-		return m.onWorkerAnnounced(value.MasterID, value.WorkerID)
-	case *blockchain.WorkerConfirmedData:
-		return m.onWorkerConfirmed(value.MasterID, value.WorkerID)
-	case *blockchain.WorkerRemovedData:
-		return m.onWorkerRemoved(value.MasterID, value.WorkerID)
-	case *blockchain.AddedToBlacklistData:
-		return m.onAddedToBlacklist(value.AdderID, value.AddeeID)
-	case *blockchain.RemovedFromBlacklistData:
-		m.onRemovedFromBlacklist(value.RemoverID, value.RemoveeID)
-	case *blockchain.ValidatorCreatedData:
-		return m.onValidatorCreated(value.ID)
-	case *blockchain.ValidatorDeletedData:
-		return m.onValidatorDeleted(value.ID)
-	case *blockchain.CertificateCreatedData:
-		return m.onCertificateCreated(value.ID)
-	}
-
-	return nil
-}
-
-func (m *DWH) onNumBenchmarksUpdated(newNumBenchmarks uint64) error {
-	if err := m.setupDB(); err != nil {
-		return fmt.Errorf("failed to setupDB after NumBenchmarksUpdated event: %v", err)
-	}
-
-	if err := m.storage.CreateIndices(m.db); err != nil {
-		return fmt.Errorf("failed to CreateIndices (onNumBenchmarksUpdated): %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onDealOpened(dealID *big.Int) error {
-	deal, err := m.blockchain.Market().GetDealInfo(m.ctx, dealID)
-	if err != nil {
-		return fmt.Errorf("failed to GetDealInfo: %v", err)
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	if deal.Status == pb.DealStatus_DEAL_CLOSED {
-		if err := m.storage.StoreStaleID(newSimpleConn(m.db), dealID, "Deal"); err != nil {
-			return fmt.Errorf("failed to StoreStaleID: %v", err)
-		}
-		m.logger.Debug("skipping inactive deal", zap.String("deal_id", dealID.String()))
-		return nil
-	}
-
-	err = m.storage.InsertDeal(conn, deal)
-	if err != nil {
-		return fmt.Errorf("failed to insertDeal: %v", err)
-	}
-
-	err = m.storage.InsertDealCondition(conn,
-		&pb.DealCondition{
-			SupplierID:  deal.SupplierID,
-			ConsumerID:  deal.ConsumerID,
-			MasterID:    deal.MasterID,
-			Duration:    deal.Duration,
-			Price:       deal.Price,
-			StartTime:   deal.StartTime,
-			EndTime:     &pb.Timestamp{},
-			TotalPayout: deal.TotalPayout,
-			DealID:      deal.Id,
-		},
-	)
-	if err != nil {
-		return fmt.Errorf("onDealOpened: failed to insertDealCondition: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onDealUpdated(dealID *big.Int) error {
-	deal, err := m.blockchain.Market().GetDealInfo(m.ctx, dealID)
-	if err != nil {
-		return fmt.Errorf("failed to GetDealInfo: %v", err)
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	// If deal is known to be stale:
-	if ok, err := m.storage.CheckStaleID(conn, dealID, "Deal"); err != nil {
-		return fmt.Errorf("failed to CheckStaleID: %v", err)
-	} else {
-		if ok {
-			m.removeStaleEntityID(dealID, "Deal")
-			return nil
-		}
-	}
-
-	if deal.Status == pb.DealStatus_DEAL_CLOSED {
-		err = m.storage.DeleteDeal(conn, deal.Id.Unwrap())
-		if err != nil {
-			return fmt.Errorf("failed to delete deal (possibly old log entry): %v", err)
-		}
-
-		if err := m.storage.DeleteOrder(conn, deal.AskID.Unwrap()); err != nil {
-			return fmt.Errorf("failed to deleteOrder: %v", err)
-		}
-		if err := m.storage.DeleteOrder(conn, deal.BidID.Unwrap()); err != nil {
-			return fmt.Errorf("failed to deleteOrder: %v", err)
-		}
-
-		return nil
-	}
-
-	if err := m.storage.UpdateDeal(conn, deal); err != nil {
-		return fmt.Errorf("failed to UpdateDeal: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onDealChangeRequestSent(eventTS uint64, changeRequestID *big.Int) error {
-	changeRequest, err := m.blockchain.Market().GetDealChangeRequestInfo(m.ctx, changeRequestID)
-	if err != nil {
-		return err
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	// If deal is known to be stale, skip.
-	if ok, err := m.storage.CheckStaleID(conn, changeRequest.DealID.Unwrap(), "Deal"); err != nil {
-		return fmt.Errorf("failed to CheckStaleID: %v", err)
-	} else {
-		if ok {
-			m.logger.Debug("skipping DealChangeRequestSent event for inactive deal")
-			return nil
-		}
-	}
-
-	if changeRequest.Status != pb.ChangeRequestStatus_REQUEST_CREATED {
-		m.logger.Info("onDealChangeRequest event points to DealChangeRequest with .Status != Created",
-			zap.String("actual_status", pb.ChangeRequestStatus_name[int32(changeRequest.Status)]))
-		return nil
-	}
-
-	// Sanity check: if more than 1 CR of one type is created for a Deal, we delete old CRs.
-	expiredChangeRequests, err := m.storage.GetDealChangeRequests(conn, changeRequest)
-	if err != nil {
-		return errors.New("failed to get (possibly) expired DealChangeRequests")
-	}
-
-	for _, expiredChangeRequest := range expiredChangeRequests {
-		err := m.storage.DeleteDealChangeRequest(conn, expiredChangeRequest.Id.Unwrap())
-		if err != nil {
-			return fmt.Errorf("failed to deleteDealChangeRequest: %v", err)
-		} else {
-			m.logger.Warn("deleted expired DealChangeRequest",
-				zap.String("id", expiredChangeRequest.Id.Unwrap().String()))
-		}
-	}
-
-	changeRequest.CreatedTS = &pb.Timestamp{Seconds: int64(eventTS)}
-	if err := m.storage.InsertDealChangeRequest(conn, changeRequest); err != nil {
-		return fmt.Errorf("failed to InsertDealChangeRequest (%s): %v", changeRequest.Id.Unwrap().String(), err)
-	}
-
-	return err
-}
-
-func (m *DWH) onDealChangeRequestUpdated(eventTS uint64, changeRequestID *big.Int) error {
-	changeRequest, err := m.blockchain.Market().GetDealChangeRequestInfo(m.ctx, changeRequestID)
-	if err != nil {
-		return err
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	// If deal is known to be stale, skip.
-	if ok, err := m.storage.CheckStaleID(conn, changeRequest.DealID.Unwrap(), "Deal"); err != nil {
-		return fmt.Errorf("failed to CheckStaleID: %v", err)
-	} else {
-		if ok {
-			m.logger.Debug("skipping DealChangeRequestUpdated event for inactive deal")
-			return nil
-		}
-	}
-
-	switch changeRequest.Status {
-	case pb.ChangeRequestStatus_REQUEST_REJECTED:
-		err := m.storage.UpdateDealChangeRequest(conn, changeRequest)
-		if err != nil {
-			return fmt.Errorf("failed to update DealChangeRequest %s: %v", changeRequest.Id.Unwrap().String(), err)
-		}
-	case pb.ChangeRequestStatus_REQUEST_ACCEPTED:
-		deal, err := m.storage.GetDealByID(conn, changeRequest.DealID.Unwrap())
-		if err != nil {
-			return fmt.Errorf("failed to storage.GetDealByID: %v", err)
-		}
-
-		deal.Deal.Duration = changeRequest.Duration
-		deal.Deal.Price = changeRequest.Price
-		if err := m.storage.UpdateDeal(conn, deal.Deal); err != nil {
-			return fmt.Errorf("failed to UpdateDeal: %v", err)
-		}
-
-		if err := m.updateDealConditionEndTime(conn, deal.GetDeal().Id, eventTS); err != nil {
-			return fmt.Errorf("failed to updateDealConditionEndTime: %v", err)
-		}
-
-		err = m.storage.InsertDealCondition(conn,
-			&pb.DealCondition{
-				SupplierID:  deal.GetDeal().SupplierID,
-				ConsumerID:  deal.GetDeal().ConsumerID,
-				MasterID:    deal.GetDeal().MasterID,
-				Duration:    changeRequest.Duration,
-				Price:       changeRequest.Price,
-				StartTime:   &pb.Timestamp{Seconds: int64(eventTS)},
-				EndTime:     &pb.Timestamp{},
-				TotalPayout: pb.NewBigIntFromInt(0),
-				DealID:      deal.GetDeal().Id,
-			},
-		)
-		if err != nil {
-			return fmt.Errorf("failed to insertDealCondition: %v", err)
-		}
-
-		err = m.storage.DeleteDealChangeRequest(conn, changeRequest.Id.Unwrap())
-		if err != nil {
-			return fmt.Errorf("failed to delete DealChangeRequest %s: %v", changeRequest.Id.Unwrap().String(), err)
-		}
-	default:
-		err := m.storage.DeleteDealChangeRequest(conn, changeRequest.Id.Unwrap())
-		if err != nil {
-			return fmt.Errorf("failed to delete DealChangeRequest %s: %v", changeRequest.Id.Unwrap().String(), err)
-		}
-	}
-
-	return nil
-}
-
-func (m *DWH) onBilled(eventTS uint64, dealID, payedAmount *big.Int) error {
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	// If deal is known to be stale, skip.
-	if ok, err := m.storage.CheckStaleID(conn, dealID, "Deal"); err != nil {
-		return fmt.Errorf("failed to CheckStaleID: %v", err)
-	} else {
-		if ok {
-			m.logger.Debug("skipping Billed event for inactive deal")
-			return nil
-		}
-	}
-
-	if err := m.updateDealPayout(conn, dealID, payedAmount, eventTS); err != nil {
-		return fmt.Errorf("failed to updateDealPayout: %v", err)
-	}
-
-	dealConditions, _, err := m.storage.GetDealConditions(conn, &pb.DealConditionsRequest{DealID: pb.NewBigInt(dealID)})
-	if err != nil {
-		return fmt.Errorf("failed to GetDealConditions (last): %v", err)
-	}
-
-	if len(dealConditions) < 1 {
-		return fmt.Errorf("no deal conditions found for deal `%s`: %v", dealID.String(), err)
-	}
-
-	err = m.storage.UpdateDealConditionPayout(conn, dealConditions[0].Id,
-		big.NewInt(0).Add(dealConditions[0].TotalPayout.Unwrap(), payedAmount))
-	if err != nil {
-		return fmt.Errorf("failed to UpdateDealConditionPayout: %v", err)
-	}
-
-	if err != nil {
-		return fmt.Errorf("insertDealPayment failed: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) updateDealPayout(conn queryConn, dealID, payedAmount *big.Int, billTS uint64) error {
-	deal, err := m.storage.GetDealByID(conn, dealID)
-	if err != nil {
-		return fmt.Errorf("failed to GetDealByID: %v", err)
-	}
-
-	newDealTotalPayout := big.NewInt(0).Add(deal.Deal.TotalPayout.Unwrap(), payedAmount)
-	err = m.storage.UpdateDealPayout(conn, dealID, newDealTotalPayout, billTS)
-	if err != nil {
-		return fmt.Errorf("failed to updateDealPayout: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onOrderPlaced(eventTS uint64, orderID *big.Int) error {
-	order, err := m.blockchain.Market().GetOrderInfo(m.ctx, orderID)
-	if err != nil {
-		return fmt.Errorf("failed to GetOrderInfo: %v", err)
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	if order.OrderStatus == pb.OrderStatus_ORDER_INACTIVE && order.DealID.IsZero() {
-		if err := m.storage.StoreStaleID(conn, orderID, "Order"); err != nil {
-			return fmt.Errorf("failed to StoreStaleID: %v", err)
-		}
-		m.logger.Debug("skipping inactive order", zap.String("order_id", orderID.String()))
-		return nil
-	}
-
-	var userID common.Address
-	if order.OrderType == pb.OrderType_ASK {
-		// For Ask orders, try to get this Author's masterID, use AuthorID if not found.
-		userID, err = m.storage.GetMasterByWorker(conn, order.GetAuthorID().Unwrap())
-		if err != nil {
-			m.logger.Warn("failed to GetMasterByWorker", zap.Error(err),
-				zap.String("author_id", order.GetAuthorID().Unwrap().Hex()))
-			userID = order.GetAuthorID().Unwrap()
-		}
-	} else {
-		userID = order.GetAuthorID().Unwrap()
-	}
-
-	profile, err := m.storage.GetProfileByID(conn, userID)
-	if err != nil {
-		certificates, _ := json.Marshal([]*pb.Certificate{})
-		profile = &pb.Profile{UserID: order.AuthorID, Certificates: string(certificates)}
-	} else {
-		if order.OrderStatus == pb.OrderStatus_ORDER_ACTIVE {
-			if err := m.updateProfileStats(conn, order.OrderType, userID, 1); err != nil {
-				return fmt.Errorf("failed to updateProfileStats: %v", err)
-			}
-		}
-	}
-
-	if order.DealID == nil {
-		order.DealID = pb.NewBigIntFromInt(0)
-	}
-
-	err = m.storage.InsertOrder(conn, &pb.DWHOrder{
-		CreatedTS:            &pb.Timestamp{Seconds: int64(eventTS)},
-		CreatorIdentityLevel: profile.IdentityLevel,
-		CreatorName:          profile.Name,
-		CreatorCountry:       profile.Country,
-		CreatorCertificates:  []byte(profile.Certificates),
-		MasterID:             pb.NewEthAddress(userID),
-		Order: &pb.Order{
-			Id:             order.Id,
-			DealID:         order.DealID,
-			OrderType:      order.OrderType,
-			OrderStatus:    order.OrderStatus,
-			AuthorID:       order.AuthorID,
-			CounterpartyID: order.CounterpartyID,
-			Duration:       order.Duration,
-			Price:          order.Price,
-			Netflags:       order.Netflags,
-			IdentityLevel:  order.IdentityLevel,
-			Blacklist:      order.Blacklist,
-			Tag:            order.Tag,
-			FrozenSum:      order.FrozenSum,
-			Benchmarks:     order.Benchmarks,
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("failed to insertOrder: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onOrderUpdated(orderID *big.Int) error {
-	marketOrder, err := m.blockchain.Market().GetOrderInfo(m.ctx, orderID)
-	if err != nil {
-		return fmt.Errorf("failed to GetOrderInfo: %v", err)
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	// If the order was known to be inactive, delete it from the list of inactive entities
-	// and skip.
-	if ok, err := m.storage.CheckStaleID(conn, orderID, "Order"); err != nil {
-		return fmt.Errorf("failed to CheckStaleID: %v", err)
-	} else {
-		if ok {
-			m.removeStaleEntityID(orderID, "Order")
-			return nil
-		}
-	}
-
-	// A situation is possible when user places an Ask order without specifying her `MasterID` (and we take
-	// `AuthorID` for `MasterID`), and afterwards the user *does* specify her master. To avoid inconsistency,
-	// we always use the user ID that was chosen in `onOrderPlaced` (i.e., the one that is already stored in DB).
-	dwhOrder, err := m.storage.GetOrderByID(conn, marketOrder.GetId().Unwrap())
-	if err != nil {
-		return fmt.Errorf("failed to GetOrderByID: %v", err)
-	}
-
-	var userID common.Address
-	if marketOrder.OrderType == pb.OrderType_ASK {
-		userID = dwhOrder.GetMasterID().Unwrap()
-	} else {
-		userID = marketOrder.GetAuthorID().Unwrap()
-	}
-
-	// If order was updated, but no deal is associated with it, delete the order.
-	if marketOrder.DealID.IsZero() {
-		if err := m.storage.DeleteOrder(conn, orderID); err != nil {
-			m.logger.Info("failed to delete Order (possibly old log entry)", zap.Error(err),
-				zap.String("order_id", orderID.String()))
-		}
-	} else {
-		// Otherwise update order status.
-		err := m.storage.UpdateOrderStatus(conn, marketOrder.Id.Unwrap(), marketOrder.OrderStatus)
-		if err != nil {
-			return fmt.Errorf("failed to updateOrderStatus (possibly old log entry): %v", err)
-		}
-	}
-
-	if dwhOrder.GetOrder().OrderStatus == pb.OrderStatus_ORDER_ACTIVE {
-		if err := m.updateProfileStats(conn, marketOrder.OrderType, userID, -1); err != nil {
-			return fmt.Errorf("failed to updateProfileStats (AuthorID: `%s`): %v", marketOrder.AuthorID.Unwrap().String(), err)
-		}
-	}
-
-	return nil
-}
-
-func (m *DWH) onWorkerAnnounced(masterID, slaveID common.Address) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if ok, err := m.storage.CheckWorkerExists(conn, masterID, slaveID); err != nil {
-		return fmt.Errorf("failed to CheckWorker: %v", err)
-	} else {
-		if ok {
-			// Worker already exists, skipping.
-			return nil
-		}
-	}
-
-	if err := m.storage.InsertWorker(conn, masterID, slaveID); err != nil {
-		return fmt.Errorf("onWorkerAnnounced failed: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onWorkerConfirmed(masterID, slaveID common.Address) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.UpdateWorker(conn, masterID, slaveID); err != nil {
-		return fmt.Errorf("onWorkerConfirmed failed: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onWorkerRemoved(masterID, slaveID common.Address) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.DeleteWorker(conn, masterID, slaveID); err != nil {
-		return fmt.Errorf("onWorkerRemoved failed: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onAddedToBlacklist(adderID, addeeID common.Address) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.InsertBlacklistEntry(conn, adderID, addeeID); err != nil {
-		return fmt.Errorf("onAddedToBlacklist failed: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onRemovedFromBlacklist(removerID, removeeID common.Address) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.DeleteBlacklistEntry(conn, removerID, removeeID); err != nil {
-		return fmt.Errorf("onRemovedFromBlacklist failed: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onValidatorCreated(validatorID common.Address) error {
-	validator, err := m.blockchain.ProfileRegistry().GetValidator(m.ctx, validatorID)
-	if err != nil {
-		return fmt.Errorf("failed to get validator `%s`: %v", validatorID.String(), err)
-	}
-
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.InsertOrUpdateValidator(conn, validator); err != nil {
-		return fmt.Errorf("failed to insertValidator: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onValidatorDeleted(validatorID common.Address) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.DeactivateValidator(conn, validatorID); err != nil {
-		return fmt.Errorf("failed to InsertOrUpdateValidator: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) onCertificateCreated(certificateID *big.Int) error {
-	certificate, err := m.blockchain.ProfileRegistry().GetCertificate(m.ctx, certificateID)
-	if err != nil {
-		return fmt.Errorf("failed to GetCertificate: %v", err)
-	}
-
-	conn, err := newTxConn(m.db, m.logger)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %v", err)
-	}
-	defer conn.Finish()
-
-	// Check if this certificate is assigned to a validator.
-	validatorLevel, err := m.blockchain.ProfileRegistry().GetValidatorLevel(m.ctx, certificate.OwnerID.Unwrap())
-	if err != nil {
-		return fmt.Errorf("failed to GetValidatorLevel: %v", err)
-	}
-	if validatorLevel != 0 {
-		// It's a validator.
-		return m.storage.UpdateValidator(
-			conn, certificate.OwnerID.Unwrap(), certificate.GetAttributeNameNormalized(), certificate.GetValue())
-	}
-
-	if err = m.storage.InsertCertificate(conn, certificate); err != nil {
-		return fmt.Errorf("failed to insertCertificate: %v", err)
-	}
-
-	if err := m.updateProfile(conn, certificate); err != nil {
-		return fmt.Errorf("failed to updateProfile: %v", err)
-	}
-
-	if err := m.updateEntitiesByProfile(conn, certificate); err != nil {
-		return fmt.Errorf("failed to updateEntitiesByProfile: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) updateProfile(conn queryConn, certificate *pb.Certificate) error {
-	_, activeAsks, err := m.storage.GetOrders(conn, &pb.OrdersRequest{
-		Type:      pb.OrderType_ASK,
-		MasterID:  certificate.OwnerID,
-		WithCount: true})
-	if err != nil {
-		return fmt.Errorf("failed to get active ASKs count: %v", err)
-	}
-
-	_, activeBids, err := m.storage.GetOrders(conn, &pb.OrdersRequest{
-		Type:      pb.OrderType_BID,
-		MasterID:  certificate.OwnerID,
-		WithCount: true})
-	if err != nil {
-		return fmt.Errorf("failed to get active BIDs count: %v", err)
-	}
-
-	certBytes, _ := json.Marshal([]*pb.Certificate{})
-	err = m.storage.InsertProfileUserID(conn, &pb.Profile{
-		UserID:       certificate.OwnerID,
-		Certificates: string(certBytes),
-		ActiveAsks:   activeAsks,
-		ActiveBids:   activeBids,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to insertProfileUserID: %v", err)
-	}
-
-	// Update distinct Profile columns.
-	switch certificate.Attribute {
-	case CertificateName, CertificateCountry:
-		err := m.storage.UpdateProfile(conn, certificate.OwnerID.Unwrap(), attributeToString[certificate.Attribute],
-			string(certificate.Value))
-		if err != nil {
-			return fmt.Errorf("failed to UpdateProfile (%s): %v err", attributeToString[certificate.Attribute], err)
-		}
-	}
-
-	// Update certificates blob.
-	certificates, err := m.storage.GetCertificates(conn, certificate.OwnerID.Unwrap())
-	if err != nil {
-		return fmt.Errorf("failed to GetCertificates: %v", err)
-	}
-
-	certificateAttrsBytes, err := json.Marshal(certificates)
-	if err != nil {
-		return fmt.Errorf("failed to marshal certificates: %v", err)
-	}
-
-	var maxIdentityLevel uint64
-	for _, certificate := range certificates {
-		if certificate.IdentityLevel > maxIdentityLevel {
-			maxIdentityLevel = certificate.IdentityLevel
-		}
-	}
-
-	err = m.storage.UpdateProfile(conn, certificate.OwnerID.Unwrap(), "Certificates", certificateAttrsBytes)
-	if err != nil {
-		return fmt.Errorf("failed to updateProfileCertificates (Certificates): %v", err)
-	}
-
-	err = m.storage.UpdateProfile(conn, certificate.OwnerID.Unwrap(), "IdentityLevel", maxIdentityLevel)
-	if err != nil {
-		return fmt.Errorf("failed to updateProfileCertificates (Level): %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) updateEntitiesByProfile(conn queryConn, certificate *pb.Certificate) error {
-	profile, err := m.storage.GetProfileByID(conn, certificate.OwnerID.Unwrap())
-	if err != nil {
-		return fmt.Errorf("failed to getProfileInfo: %v", err)
-	}
-
-	if err := m.storage.UpdateOrders(conn, profile); err != nil {
-		return fmt.Errorf("failed to updateOrders: %v", err)
-	}
-
-	if err = m.storage.UpdateDealsSupplier(conn, profile); err != nil {
-		return fmt.Errorf("failed to updateDealsSupplier: %v", err)
-	}
-
-	err = m.storage.UpdateDealsConsumer(conn, profile)
-	if err != nil {
-		return fmt.Errorf("failed to updateDealsConsumer: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) updateProfileStats(conn queryConn, orderType pb.OrderType, userID common.Address, update int) error {
-	var field string
-	if orderType == pb.OrderType_ASK {
-		field = "ActiveAsks"
-	} else {
-		field = "ActiveBids"
-	}
-
-	if err := m.storage.UpdateProfileStats(conn, userID, field, update); err != nil {
-		return fmt.Errorf("failed to UpdateProfileStats: %v", err)
-	}
-
-	return nil
-}
-
-// coldStart waits till last seen block number gets to `w.cfg.ColdStart.UpToBlock` and then tries to create indices.
-func (m *DWH) coldStart() error {
-	ticker := time.NewTicker(time.Second * 5)
-	defer ticker.Stop()
-
-	targetBlock, err := m.blockchain.Events().GetLastBlock(m.ctx)
-	if err != nil {
-		return fmt.Errorf("failed to GetLastBlock: %v", err)
-	}
-	var retries = 5
-	for {
-		select {
-		case <-m.ctx.Done():
-			m.logger.Info("stopped coldStart routine")
-			return nil
-		case <-ticker.C:
-			targetBlockReached, err := m.maybeCreateIndices(targetBlock)
-			if err != nil {
-				if retries == 0 {
-					m.logger.Warn("failed to CreateIndices, exiting")
-					return err
-				}
-				retries--
-				m.logger.Warn("failed to CreateIndices, retrying", zap.Int("retries_left", retries))
-			}
-			if targetBlockReached {
-				m.logger.Info("CreateIndices success")
-				return nil
-			}
-		}
-	}
-}
-
-func (m *DWH) maybeCreateIndices(targetBlock uint64) (targetBlockReached bool, err error) {
-	lastEvent, err := m.getLastEvent()
-	if err != nil {
-		return false, err
-	}
-
-	m.logger.Info("current block (waiting to CreateIndices)", zap.Uint64("block_number", lastEvent.BlockNumber))
-	if lastEvent.BlockNumber >= targetBlock {
-		if err := m.storage.CreateIndices(m.db); err != nil {
-			return false, err
-		}
-		return true, nil
-	}
-
-	return false, nil
-}
-
-func (m *DWH) getLastEvent() (*blockchain.Event, error) {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	return m.storage.GetLastEvent(conn)
-}
-
-func (m *DWH) updateLastEvent(event *blockchain.Event) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.UpdateLastEvent(conn, event); err != nil {
-		return fmt.Errorf("failed to updateLastEvent: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) insertLastEvent(event *blockchain.Event) error {
-	conn := newSimpleConn(m.db)
-	defer conn.Finish()
-
-	if err := m.storage.InsertLastEvent(conn, event); err != nil {
-		return fmt.Errorf("failed to updateLastEvent: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) updateDealConditionEndTime(conn queryConn, dealID *pb.BigInt, eventTS uint64) error {
-	dealConditions, _, err := m.storage.GetDealConditions(conn, &pb.DealConditionsRequest{DealID: dealID})
-	if err != nil {
-		return fmt.Errorf("failed to getDealConditions: %v", err)
-	}
-
-	dealCondition := dealConditions[0]
-	if err := m.storage.UpdateDealConditionEndTime(conn, dealCondition.Id, eventTS); err != nil {
-		return fmt.Errorf("failed to update DealCondition: %v", err)
-	}
-
-	return nil
-}
-
-func (m *DWH) removeStaleEntityID(id *big.Int, entity string) error {
-	m.logger.Debug("removing stale entity from cache", zap.String("entity", entity), zap.String("id", id.String()))
-	if err := m.storage.RemoveStaleID(newSimpleConn(m.db), id, entity); err != nil {
-		return fmt.Errorf("failed to RemoveStaleID (%s %s): %v", entity, id.String(), err)
-	}
-
-	return nil
-}
-
-func (m *DWH) saveLastEvent() {
-	if err := m.updateLastEvent(m.lastEvent); err != nil {
-		m.logger.Warn("failed to updateLastEvent", zap.Error(err))
-	}
-}
-
-type eventsDispatcher struct {
-	logger                    *zap.Logger
-	NumBenchmarksUpdated      []*blockchain.Event
-	ValidatorsCreated         []*blockchain.Event
-	CertificatesCreated       []*blockchain.Event
-	OrdersOpened              []*blockchain.Event
-	DealsOpened               []*blockchain.Event
-	DealChangeRequestsSent    []*blockchain.Event
-	DealChangeRequestsUpdated []*blockchain.Event
-	Billed                    []*blockchain.Event
-	OrdersClosed              []*blockchain.Event
-	DealsClosed               []*blockchain.Event
-	ValidatorsDeleted         []*blockchain.Event
-	AddedToBlacklist          []*blockchain.Event
-	RemovedFromBlacklist      []*blockchain.Event
-	WorkersAnnounced          []*blockchain.Event
-	WorkersConfirmed          []*blockchain.Event
-	WorkersRemoved            []*blockchain.Event
-	Other                     []*blockchain.Event
-}
-
-func newEventDispatcher(logger *zap.Logger) *eventsDispatcher {
-	return &eventsDispatcher{logger: logger}
-}
-
-func (m *eventsDispatcher) Add(event *blockchain.Event) {
-	switch data := event.Data.(type) {
-	case *blockchain.NumBenchmarksUpdatedData:
-		m.NumBenchmarksUpdated = append(m.NumBenchmarksUpdated, event)
-	case *blockchain.ValidatorCreatedData:
-		m.ValidatorsCreated = append(m.ValidatorsCreated, event)
-	case *blockchain.ValidatorDeletedData:
-		m.ValidatorsDeleted = append(m.ValidatorsDeleted, event)
-	case *blockchain.CertificateCreatedData:
-		m.CertificatesCreated = append(m.CertificatesCreated, event)
-	case *blockchain.DealOpenedData:
-		m.DealsOpened = append(m.DealsOpened, event)
-	case *blockchain.DealUpdatedData:
-		m.DealsClosed = append(m.DealsClosed, event)
-	case *blockchain.OrderPlacedData:
-		m.OrdersOpened = append(m.OrdersOpened, event)
-	case *blockchain.OrderUpdatedData:
-		m.OrdersClosed = append(m.OrdersClosed, event)
-	case *blockchain.DealChangeRequestSentData:
-		m.DealChangeRequestsSent = append(m.DealChangeRequestsSent, event)
-	case *blockchain.DealChangeRequestUpdatedData:
-		m.DealChangeRequestsUpdated = append(m.DealChangeRequestsUpdated, event)
-	case *blockchain.BilledData:
-		m.Billed = append(m.Billed, event)
-	case *blockchain.AddedToBlacklistData:
-		m.AddedToBlacklist = append(m.AddedToBlacklist, event)
-	case *blockchain.RemovedFromBlacklistData:
-		m.RemovedFromBlacklist = append(m.RemovedFromBlacklist, event)
-	case *blockchain.WorkerAnnouncedData:
-		m.WorkersAnnounced = append(m.WorkersAnnounced, event)
-	case *blockchain.WorkerConfirmedData:
-		m.WorkersConfirmed = append(m.WorkersConfirmed, event)
-	case *blockchain.WorkerRemovedData:
-		m.WorkersRemoved = append(m.WorkersRemoved, event)
-	case *blockchain.ErrorData:
-		m.logger.Warn("received error from events channel", zap.Error(data.Err), zap.String("topic", data.Topic))
-	default:
-		m.Other = append(m.Other, event)
-	}
 }
